Refactor status command into SRP modules
